+++ core/amazon/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "amazon-product",
+  "name": "amazon",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
